+++ src/main/kotlin/com/insyncwithfoo/pyright/runner/Runner.kt
@@ -3,37 +3,61 @@ package com.insyncwithfoo.pyright.runner
 import com.insyncwithfoo.pyright.PyrightOutput
 import com.intellij.execution.RunCanceledByUserException
 import com.intellij.openapi.diagnostic.Logger
+import com.intellij.openapi.project.Project
+import kotlinx.serialization.Serializable
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 
 
-class PyrightRunner(private val command: PyrightCommand) {
+@Serializable
+private data class ExceptionInfo(
+    val type: String,
+    val stdout: String,
+    val stderr: String,
+    val message: String
+)
+
+
+private fun ExceptionInfo(exception: PyrightException): ExceptionInfo {
+    return ExceptionInfo(
+        exception::class.simpleName!!,
+        exception.stdout,
+        exception.stderr,
+        exception.message!!
+    )
+}
+
+
+class PyrightRunner(project: Project) {
     
-    fun run(): PyrightOutput? {
+    private val notifier = Notifier(project)
+    
+    fun run(command: PyrightCommand): PyrightOutput? {
         val serializedCommand = Json.encodeToString(command)
         LOGGER.info("Running: $serializedCommand")
         
         return try {
-            runAndLogOutput()
+            runAndLogOutput(command)
         } catch (_: RunCanceledByUserException) {
             null
-        } catch (exception: FatalException) {
-            ERROR_REPORTER.reportException(exception)
-            null
-        } catch (exception: InvalidParametersException) {
-            ERROR_REPORTER.reportException(exception)
-            null
+        } catch (exception: PyrightException) {
+            report(exception)
+            
+            when (exception) {
+                is InvalidConfigurationsException -> parseOutput(exception.stdout)
+                is FatalException, is InvalidParametersException -> null
+            }
         }
     }
     
-    private fun runAndLogOutput(): PyrightOutput {
+    private fun runAndLogOutput(command: PyrightCommand): PyrightOutput {
         val output = command.run()
-        val parsed = parseOutput(output).also { logOutput(it) }
+        val parsed = parseOutput(output).also { logMinified(it) }
         
         return parsed
     }
     
-    private fun logOutput(parsedOutput: PyrightOutput) {
+    private fun logMinified(parsedOutput: PyrightOutput) {
         val minified = Json.encodeToString(parsedOutput)
         LOGGER.info("Output: $minified")
     }
@@ -41,9 +65,17 @@ class PyrightRunner(private val command: PyrightCommand) {
     private fun parseOutput(response: String) =
         Json.decodeFromString<PyrightOutput>(response)
     
+    private fun report(exception: PyrightException) {
+        val exceptionInfo = Json.encodeToString(ExceptionInfo(exception))
+        
+        LOGGER.warn(exception)
+        LOGGER.info("Exception properties: $exceptionInfo")
+        
+        notifier.notify(exception)
+    }
+    
     companion object {
         private val LOGGER = Logger.getInstance(PyrightRunner::class.java)
-        private val ERROR_REPORTER = PyrightErrorReporter(LOGGER)
     }
     
 }